% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVaR.outputs.R
\name{cvar.es}
\alias{cvar.es}
\title{CVAR and Expected Shortfall loss calculation}
\usage{
cvar.es(port_loss, alpha = c(0.99, 0.995, 0.999))
}
\arguments{
\item{port_loss}{vector of scenario total portfolio loss amounts}

\item{alpha}{vector of confidence levels for VAR and ES calculation. Defaults to (99%, 99.5%, 99.9%)}
}
\value{
list with a vector VAR and ES with a value relating to each alpha
}
\description{
Function to calculate Value-at-Risk and Expected Shortfall
}
\examples{
cvar.es(c(0, 1000, 5000), c(0.95, 0.99))
}
\keyword{CVAR}
\keyword{ES}
\keyword{portfolio}
