% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVaR.outputs.R
\name{cvar.out}
\alias{cvar.out}
\title{Print and graph outputs of the CVAR model}
\usage{
cvar.out(portfolio, scen_PD, port_loss, VAR, ExSf)
}
\arguments{
\item{portfolio}{input portfolio}

\item{scen_PD}{mean default rate over all scenarios}

\item{port_loss}{vector of losses across all scenarios}

\item{VAR}{value at risk}

\item{ExSf}{Expected Shortfall}
}
\value{
prints outputs and histogram
}
\description{
Function to generate printed and graph outputs of the CVAR model
}
\examples{
cvar.out(port, 0.02, port_loss, VAR, ES)
}
\keyword{cvar}
\keyword{list}
\keyword{portfolio}
