% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVaR-Portfolio.R
\name{cvar.port.b2mort}
\alias{cvar.port.b2mort}
\title{Basel II homogeneous mortgage loan portfolio generator}
\usage{
cvar.port.b2mort(NC = 100)
}
\arguments{
\item{NC}{Number of counterparties. Default is 100}
}
\description{
Function to generate a homogeneous mortgage loan portfolio of specified size. For use in benchmarking CVaR model outputs vs Basel II capital calculations. Outputs a list of: portfolio dataframe, covar matrix and R-Squared correlations
}
\examples{
cvar.port.gen()
}
\keyword{cvar}
\keyword{list}
\keyword{portfolio}
