% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVaR-Portfolio.R
\name{cvar.port.gen}
\alias{cvar.port.gen}
\title{Generic loan portfolio generator}
\usage{
cvar.port.gen(NC = 100)
}
\arguments{
\item{NC}{Number of counterparties. Default is 100.}
}
\value{
Outputs a list of 3 elements: portfolio dataframe, covar matrix and R-Squared correlations vector
}
\description{
Function to generate a randomised generic loan portfolio of specified size.
}
\examples{
cvar.port.gen(125)
}
\keyword{cvar}
\keyword{list}
\keyword{portfolio}
